Prototype Code for plotting Fidelity
In this notebook, GHZ states are prepared (Static, Dynamic Circuits) with number of qubits 3 to 8 (3 copies of each) therefore total 36 circuits have been run in order -
[Static 1, Dynamic 1, Static 1, Dynamic 1, Static 1, Dynamic 1, Static 2, Dynamic 2, Static 2, Dynamic 2, Static 2, Dynamic 2,  ......... ].
We also simulated 4 hardwares, compared their results with ideal simulation and plotted the mean Fidelities with their standard deviation. We can compare performance of Dynamic Circuit with Static Circuit with this.