Prototype code for fidelity comparison
The first circuit is the static circuit, dynamic_1 is the dynamic circuit and dynamic_2 is the dynamic circuit created using plugin which is similar to the static due to no qubit reset. A kernel restart is required before running otherwise, the classical register gets renamed. The classical registers of dynamic_2 need to be defined separately as attributes. The optimization is 1 in this case.