2nd test of GHZ 3 to 19 qubits on ibm_brisbane
The first circuit is the static circuit, dynamic_1 is the dynamic circuit and dynamic_2 is the dynamic circuit created using plugin. Restart of the kernel is required before running otherwise classical register gets renamed. The classical registers of dynamic_2 needs to be defined separately as attributes. ibm_brisbane is taken as the real hardware. The qiskit runtime usage along with other metrics of the job is present.